backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
openai
python-dotenv